# contracts/openapi/agent-studio-api.yaml
openapi: 3.1.0
info:
  title: Agent Studio API
  version: 0.1.0
  description: >-
    Task API contract. Authentication and authorization are implemented by platform-api.
    Screenshot tasks require an explicit application baseUrl in parameters; login profile URLs
    are used only for authentication and are never treated as the application entry point.
paths:
...
  /v1/tasks/{taskId}/events:
    get:
      operationId: streamTaskEvents
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - name: Last-Event-ID
          in: header
          required: false
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Server-sent task events, ordered by per-task sequence.
          content:
            text/event-stream:
              schema: { $ref: '#/components/schemas/TaskEvent' }
...
components:
  parameters:
    TaskId: { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
...
  schemas:
...
    TaskStatus: { type: string, enum: [QUEUED, RUNNING, PAUSED, WAITING_FOR_APPROVAL, SUCCEEDED, FAILED, CANCELED] }
...
    TaskEvent:
      type: object
      required: [taskId, sequence, status, occurredAt]
      properties:
        taskId: { type: string, format: uuid }
        sequence: { type: integer, minimum: 1 }
        status: { $ref: '#/components/schemas/TaskStatus' }
        occurredAt: { type: string, format: date-time }
        resultReference: { type: [string, 'null'] }
        failureCode: { type: [string, 'null'] }